+++ smawg/__init__.py
@@ -3,11 +3,14 @@
 See https://github.com/expurple/smawg for more info about the project.
 """
 
+from smawg._common import (
+    Ability, AbstractRules, Combo, GameState,
+    Player, Race, Region, RulesViolation
+)
 from smawg._engine import Game, Hooks, InvalidAssets, validate
-from smawg.common import Ability, Combo, Player, Race, Region, RulesViolation
 
 __all__ = [
     "Game", "Hooks", "InvalidAssets", "validate",
-    "Ability", "Combo", "Player", "Race", "Region",
-    "RulesViolation"
+    "Ability", "AbstractRules", "Combo", "GameState",
+    "Player", "Race", "Region", "RulesViolation"
 ]